utils: Extract pretty printers into a header
Can be easily reused elsewhere.
