+++ compaction/compaction.hh
@@ -14,6 +14,7 @@
 #include "gc_clock.hh"
 #include "compaction_weight_registration.hh"
 #include "utils/UUID.hh"
+#include "utils/pretty_printers.hh"
 #include "table_state.hh"
 #include <seastar/core/thread.hh>
 #include <seastar/core/abort_source.hh>
@@ -23,21 +24,6 @@ using namespace compaction;
 namespace sstables {
 
 bool is_eligible_for_compaction(const sstables::shared_sstable& sst) noexcept;
-
-class pretty_printed_data_size {
-    uint64_t _size;
-public:
-    pretty_printed_data_size(uint64_t size) : _size(size) {}
-    friend std::ostream& operator<<(std::ostream&, pretty_printed_data_size);
-};
-
-class pretty_printed_throughput {
-    uint64_t _size;
-    std::chrono::duration<float> _duration;
-public:
-    pretty_printed_throughput(uint64_t size, std::chrono::duration<float> dur) : _size(size), _duration(std::move(dur)) {}
-    friend std::ostream& operator<<(std::ostream&, pretty_printed_throughput);
-};
 
 // Return the name of the compaction type
 // as used over the REST api, e.g. "COMPACTION" or "CLEANUP".
